Finished Eval function, fixed name-related errors
Finished the evaluation function of the model and, after attempting to run the program, changed some typos in the code. Had to uninstall pytorch and manually reinstall via the conda command to match with torch versioning - the Anaconda plugin is behind updates a decent amount compared to running the command.